Update mfrc630
In write_FIFO function by this structure when we want to read a card, the IRQ0 returned 0x2A, which is indicate the ErrIRQ, because data do not write to FIFO  correctly. to overcome the issue I modified the function structure.
In write FIFO function we must write REG_FIFODATA before data in every 1 byte of data.